Refactor - determineCoinValueBasedOnWeightAndSizeByDiameter
I decided to refactor this method a bit. I wanted it to be more readible. I moved what was being evaluted in the if, else if statements from 1,2,3 to more readable field constants. I also extracted the repeated statements and put them into a single method. All the tests are still working fine and are in a Green State.