Refactor - made pressChipButton method more generic
I decided to make the pressChipButton method a bit more generic. If I would have kept going down the same route as the PressChipButton, well I would have to create 2 other methods for Candy and Cola so it made sense to make the chip method more generic which is now a pressButtonMethod that takes in the item name.