Refactor - made doWeHaveChipsInStock method more generic
I decided to make the doWeHaveChipsInStock method more generic, the method name changed to doWeHaveItemInStock. I decided to make it more generic because I did not want to write the same method for Cola and Candy so this new method will be great to handle those items once I add them in.